containers/custom_job/pyproject.toml
[tool.poetry]
name = "custom_job"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-aiplatform = "^1.16.0"
pytz = "^2022.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
